refactor(main): extract host construction into host/createHostWindow
